Cambio a int

+++ POOConcepts/Taller1POO/Time.cs
@@ -66,20 +66,20 @@ class Time
     // Método IsOtherDay
     public bool IsOtherDay(Time other)
     {
-        long totalMilliseconds = this.ToMilliseconds() + other.ToMilliseconds();
+        int totalMilliseconds = this.ToMilliseconds() + other.ToMilliseconds();
         return totalMilliseconds >= 86400000; // 86400000 ms = 24 horas
     }
 
     // Método Add
     public Time Add(Time other)
     {
-        long totalMilliseconds = this.ToMilliseconds() + other.ToMilliseconds();
-        long hours = (totalMilliseconds / 3600000) % 24;
-        long minutes = (totalMilliseconds % 3600000) / 60000;
-        long seconds = (totalMilliseconds % 60000) / 1000;
-        long milliseconds = totalMilliseconds % 1000;
+        int totalMilliseconds = this.ToMilliseconds() + other.ToMilliseconds();
+        int hours = (totalMilliseconds / 3600000) % 24;
+        int minutes = (totalMilliseconds % 3600000) / 60000;
+        int seconds = (totalMilliseconds % 60000) / 1000;
+        int milliseconds = totalMilliseconds % 1000;
 
-        return new Time((int)hours, (int)minutes, (int)seconds, (int)milliseconds);
+        return new Time(hours, minutes, seconds, milliseconds);
     }
 
 }
